backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo[srv]==3.12
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Pillow==10.0.1
email-validator==2.1.0
bcrypt==4.0.1
python-multipart==0.0.6
cloudinary==1.36.0
gunicorn==21.2.0
